Conflicts:
	src/main/java/org/elasticsearch/action/search/type/TransportSearchQueryThenFetchAction.java
	src/main/java/org/elasticsearch/search/SearchService.java
	src/main/java/org/elasticsearch/search/internal/SearchContext.java
	src/main/java/org/elasticsearch/search/query/QueryPhase.java

+++ src/main/java/org/elasticsearch/search/facet/filter/FilterFacetExecutor.java
@@ -1,11 +1,11 @@
 /*
- * Licensed to ElasticSearch and Shay Banon under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership. ElasticSearch licenses this
- * file to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
+ * Licensed to Elasticsearch under one or more contributor
+ * license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright
+ * ownership. Elasticsearch licenses this file to you under
+ * the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
  *
  *    http://www.apache.org/licenses/LICENSE-2.0
  *
@@ -40,7 +40,7 @@ public class FilterFacetExecutor extends FacetExecutor {
 
     private final Filter filter;
 
-    int count = -1;
+    long count = -1;
 
     public FilterFacetExecutor(Filter filter) {
         this.filter = filter;
@@ -82,7 +82,7 @@ public class FilterFacetExecutor extends FacetExecutor {
 
     class Collector extends FacetExecutor.Collector {
 
-        private int count = 0;
+        private long count = 0;
         private Bits bits;
 
         @Override
